more refactoring of map.helper into functions, moving things around to separate files
